+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - 修复作者的大小写。 / Fix case of creators.
 
+### Fixed
+
+- 菜单中的大小写。/ Fix typo.
+
 ## [1.0.7] - 2023-08-07
 
 ### Added